fix(review): remediate skill:checklist-216:DESIGN-OPTIMIZATION-TRADEOFF

+++ src/engine/export_prototypes/VectorizedPrefixSlicer.h
@@ -42,7 +42,7 @@ class VectorizedPrefixTable {
     // Stores a prefix aligned to 16 bytes with its length.
   alignas(16) struct PrefixEntry {
     alignas(16) char data[48];
-    size_t length;
+    uint8_t length;
   };
 
  private:
